+++ lib/cgo/tests/check_coin.block.c
@@ -29,21 +29,23 @@ Transactions__Handle makeTestTransactions()
 GoUint32 makeNewBlock(cipher__SHA256* uxHash, Block__Handle* newBlock)
 {
     GoUint32 result;
-    cipher__SHA256 bodyhash;
-    BlockBody__Handle block;
-    Transactions__Handle transactions = makeTestTransactions();
+    cipher__SHA256 bodyhash = "";
+    BlockBody__Handle block = 0;
+    Transactions__Handle transactions = 0;
+    transactions = makeTestTransactions();
 
     result = SKY_coin_NewEmptyBlock(transactions, &block);
     ck_assert_msg(result == SKY_OK, "SKY_coin_NewEmptyBlock failed");
     registerHandleClose(block);
-    coin__Block* pBlock;
-    result = SKY_coin_GetBlockObject(block, &pBlock);
-    ck_assert_msg(result == SKY_OK, "SKY_coin_Get_Block_Object failed");
+    BlockHeader__Handle pBlockHeader = 0;
+    result = SKY_coin_Block_GetBlockHeader(block, &pBlockHeader);
+    ck_assert_msg(result == SKY_OK, "SKY_coin_Block_GetBlockHeader failed");
 
-    pBlock->Head.Version = 0x02;
-    pBlock->Head.Time = 100;
-    pBlock->Head.BkSeq = 0;
-    pBlock->Head.Fee = 10;
+    SKY_coin_BlockHeader_SetTime(pBlockHeader, 100);
+    SKY_coin_BlockHeader_SetBkSeq(pBlockHeader, 0);
+    SKY_coin_BlockHeader_SetVersion(pBlockHeader, 0x02);
+    SKY_coin_BlockHeader_SetFee(pBlockHeader, 10);
+
     BlockBody__Handle body = 0;
     result = SKY_coin_GetBlockBody(block, &body);
     ck_assert_msg(result == SKY_OK, "SKY_coin_Get_Block_Body failed");
@@ -51,7 +53,7 @@ GoUint32 makeNewBlock(cipher__SHA256* uxHash, Block__Handle* newBlock)
     ck_assert_msg(result == SKY_OK, "SKY_coin_BlockBody_Hash failed");
     FeeCalculator zf = {zeroFeeCalculator, NULL};
     result = SKY_coin_NewBlock(block, 100 + 200, uxHash, transactions, &zf, newBlock);
-    ck_assert_msg(result == SKY_OK, "SKY_coin_NewBlock failed");
+    ck_assert_int_eq(result, SKY_OK);
     registerHandleClose(*newBlock);
     return result;
 }
@@ -76,16 +78,16 @@ START_TEST(TestNewBlock)
     ck_assert_msg(result == SKY_OK, "SKY_coin_NewEmptyBlock failed");
     registerHandleClose(prevBlock);
     coin__Block* pBlock;
-    result = SKY_coin_GetBlockObject(prevBlock, &pPrevBlock);
-    ck_assert_msg(result == SKY_OK, "SKY_coin_GetBlockObject failed");
+    BlockHeader__Handle pPrevBlockHeader = 0;
+    result = SKY_coin_Block_GetBlockHeader(prevBlock, &pPrevBlockHeader);
+    ck_assert_msg(result == SKY_OK, "SKY_coin_Block_GetBlockHeader failed");
 
-    pPrevBlock->Head.Version = 0x02;
-    pPrevBlock->Head.Time = 100;
-    pPrevBlock->Head.BkSeq = 98;
+    SKY_coin_BlockHeader_SetTime(pPrevBlockHeader, 100);
+    SKY_coin_BlockHeader_SetBkSeq(pPrevBlockHeader, 98);
 
     GoUint8 buffer_slice[1024];
     GoSlice slice = {buffer_slice, 0, 1024};
-    cipher__SHA256 hash;
+    cipher__SHA256 hash = "";
 
     randBytes(&slice, 128);
     registerMemCleanup(slice.data);
@@ -113,15 +115,26 @@ START_TEST(TestNewBlock)
     result = SKY_coin_NewBlock(prevBlock, currentTime, &hash, transactions, &f121, &newBlock);
     ck_assert_msg(result == SKY_OK, "SKY_coin_NewBlock failed");
     registerHandleClose(newBlock);
-    result = SKY_coin_GetBlockObject(newBlock, &pNewBlock);
-    ck_assert_msg(result == SKY_OK, "SKY_coin_GetBlockObject failed");
-    coin__Transactions* pTransactions = NULL;
-    SKY_coin_GetTransactionsObject(transactions, &pTransactions);
-    ck_assert(isTransactionsEq(&pNewBlock->Body.Transactions, pTransactions));
-    ck_assert(pNewBlock->Head.Fee == fee * (GoUint64)(pTransactions->len));
-    ck_assert(pNewBlock->Head.Time == currentTime);
-    ck_assert(pNewBlock->Head.BkSeq == (pPrevBlock->Head.BkSeq + 1));
-    ck_assert(isU8Eq(pNewBlock->Head.UxHash, hash, sizeof(cipher__SHA256)));
+    BlockHeader__Handle pnewBlockHeader = 0;
+    result = SKY_coin_Block_GetBlockHeader(newBlock, &pnewBlockHeader);
+    GoUint64 pHeadFee;
+    GoUint64 pHeadTime;
+    GoUint64 pHeadBkSeq;
+    SKY_coin_BlockHeader_Fee(pnewBlockHeader, &pHeadFee);
+    SKY_coin_BlockHeader_Time(pnewBlockHeader, &pHeadTime);
+    SKY_coin_BlockHeader_BkSeq(pnewBlockHeader, &pHeadBkSeq);
+    GoInt pTransLength;
+    result = SKY_coin_Transactions_Length(transactions, &pTransLength);
+    ck_assert_int_eq(pHeadFee, (fee * pTransLength));
+    ck_assert_int_eq(pHeadTime, currentTime);
+    pPrevBlockHeader = 0;
+    result = SKY_coin_Block_GetBlockHeader(prevBlock, &pPrevBlockHeader);
+    GoUint64 pPrevBkSeq;
+    SKY_coin_BlockHeader_BkSeq(pPrevBlockHeader, &pPrevBkSeq);
+    ck_assert_int_eq(pHeadBkSeq, pPrevBkSeq + 1);
+    cipher__SHA256 pnewuxHash = " ";
+    SKY_coin_BlockHeader_UxHash(pnewBlockHeader, &pnewuxHash);
+    ck_assert(isU8Eq(pnewuxHash, hash, sizeof(cipher__SHA256)));
 }
 END_TEST
 
@@ -129,26 +142,26 @@ START_TEST(TestBlockHashHeader)
 {
     printf("Load TestBlockHashHeader\n");
     GoUint32_ result;
-    Block__Handle block = 0;
-    coin__Block* pBlock = NULL;
     GoUint8 buffer_slice[1024];
     GoSlice slice = {buffer_slice, 0, 1024};
-    cipher__SHA256 hash;
-
+    cipher__SHA256 hash = "";
     randBytes(&slice, 128);
-    registerMemCleanup(slice.data);
     result = SKY_cipher_SumSHA256(slice, &hash);
     ck_assert_msg(result == SKY_OK, "SKY_cipher_SumSHA256 failed");
+    Block__Handle block = 0;
+    coin__Block* pBlock = NULL;
     result = makeNewBlock(&hash, &block);
     ck_assert_msg(result == SKY_OK, "makeNewBlock failed");
-    result = SKY_coin_GetBlockObject(block, &pBlock);
-    ck_assert_msg(result == SKY_OK, "SKY_coin_GetBlockObject failed, block handle : %d", block);
 
     cipher__SHA256 hash1 = "";
     cipher__SHA256 hash2 = "";
     result = SKY_coin_Block_HashHeader(block, &hash1);
     ck_assert_msg(result == SKY_OK, "SKY_coin_Block_HashHeader failed");
-    result = SKY_coin_BlockHeader_Hash(&pBlock->Head, &hash2);
+    BlockHeader__Handle blockheader = 0;
+    coin__BlockHeader* pblockheader = NULL;
+    result = SKY_coin_Block_GetBlockHeader(block, &blockheader);
+    ck_assert_msg(result == SKY_OK, "SKY_coin_Block_GetBlockHeader failed");
+    result = SKY_coin_BlockHeader_Hash(blockheader, &hash2);
     ck_assert_msg(result == SKY_OK, "SKY_coin_BlockHeader_Hash failed");
     ck_assert(isU8Eq(hash1, hash2, sizeof(cipher__SHA256)));
     strcpy(hash2, "");
@@ -171,13 +184,14 @@ START_TEST(TestBlockHashBody)
     result = makeNewBlock(&hash, &block);
     ck_assert_msg(result == SKY_OK, "makeNewBlock failed");
 
-    cipher__SHA256 hash1, hash2;
+    cipher__SHA256 hash1 = "";
+    cipher__SHA256 hash2 = "";
     result = SKY_coin_Block_HashBody(block, &hash1);
-    ck_assert_msg(result == SKY_OK, "SKY_coin_BlockBody_Hash failed");
-    BlockBody__Handle blockBody = 0;
-    result = SKY_coin_GetBlockBody(block, &blockBody);
+    ck_assert_msg(result == SKY_OK, "SKY_coin_Block_HashBody failed");
+    BlockBody__Handle blockbody = 0;
+    result = SKY_coin_GetBlockBody(block, &blockbody);
     ck_assert_msg(result == SKY_OK, "SKY_coin_GetBlockBody failed");
-    result = SKY_coin_BlockBody_Hash(blockBody, &hash2);
+    result = SKY_coin_BlockBody_Hash(blockbody, &hash2);
     ck_assert_msg(result == SKY_OK, "SKY_coin_BlockBody_Hash failed");
     ck_assert(isU8Eq(hash1, hash2, sizeof(cipher__SHA256)));
 }
@@ -193,34 +207,69 @@ START_TEST(TestNewGenesisBlock)
     GoUint64 genCoins = 1000 * 1000 * 1000;
     GoUint64 genCoinHours = 1000 * 1000;
     Block__Handle block = 0;
-    coin__Block* pBlock = NULL;
 
     GoUint32_ result = makeKeysAndAddress(&pubkey, &seckey, &address);
     ck_assert_msg(result == SKY_OK, "makeKeysAndAddress failed");
     result = SKY_coin_NewGenesisBlock(&address, genCoins, genTime, &block);
     ck_assert_msg(result == SKY_OK, "SKY_coin_NewGenesisBlock failed");
-    result = SKY_coin_GetBlockObject(block, &pBlock);
-    ck_assert_msg(result == SKY_OK, "SKY_coin_GetBlockObject failed");
 
-    cipher__SHA256 nullHash;
-    memset(&nullHash, 0, sizeof(cipher__SHA256));
-    ck_assert(isU8Eq(nullHash, pBlock->Head.PrevHash, sizeof(cipher__SHA256)));
-    ck_assert(genTime == pBlock->Head.Time);
-    ck_assert(0 == pBlock->Head.BkSeq);
-    ck_assert(0 == pBlock->Head.Version);
-    ck_assert(0 == pBlock->Head.Fee);
-    ck_assert(isU8Eq(nullHash, pBlock->Head.UxHash, sizeof(cipher__SHA256)));
+    cipher__SHA256 nullHash = "";
+    BlockHeader__Handle blockheader = 0;
+    coin__BlockHeader* pblockheader = NULL;
+    result = SKY_coin_Block_GetBlockHeader(block, &blockheader);
+    cipher__SHA256 pBlockPrevHash = "";
+    result = SKY_coin_BlockHeader_PrevHash(blockheader, &pBlockPrevHash);
+    ck_assert_msg(result == SKY_OK, "SKY_coin_BlockHeader_PrevHash failed");
+    ck_assert(isU8Eq(nullHash, pBlockPrevHash, sizeof(cipher__SHA256)));
 
-    ck_assert(1 == pBlock->Body.Transactions.len);
-    coin__Transaction* ptransaction = (coin__Transaction*)pBlock->Body.Transactions.data;
-    ck_assert(0 == ptransaction->In.len);
-    ck_assert(0 == ptransaction->Sigs.len);
-    ck_assert(1 == ptransaction->Out.len);
+    GoUint64 pBlockTime;
+    GoUint64 pBlockBkSeq;
+    GoUint32 pBlockVersion;
+    GoUint64 pBlockFee;
+    result = SKY_coin_BlockHeader_Time(blockheader, &pBlockTime);
+    ck_assert_msg(result == SKY_OK, "SKY_coin_BlockHeader_Time failed");
+    ck_assert(genTime == pBlockTime);
+    result = SKY_coin_BlockHeader_BkSeq(blockheader, &pBlockBkSeq);
+    ck_assert_msg(result == SKY_OK, "SKY_coin_BlockHeader_BkSeq failed");
+    ck_assert(0 == pBlockBkSeq);
+    result = SKY_coin_BlockHeader_Version(blockheader, &pBlockVersion);
+    ck_assert(result == SKY_OK);
+    ck_assert(0 == pBlockVersion);
+    result = SKY_coin_BlockHeader_Fee(blockheader, &pBlockFee);
+    ck_assert(result == SKY_OK);
+    ck_assert(0 == pBlockFee);
+    cipher__SHA256 pBlockUxHash = " ";
+    result = SKY_coin_BlockHeader_UxHash(blockheader, &pBlockUxHash);
+    ck_assert(result == SKY_OK);
+    ck_assert(isU8Eq(nullHash, pBlockUxHash, sizeof(cipher__SHA256)));
 
-    coin__TransactionOutput* poutput = (coin__TransactionOutput*)ptransaction->Out.data;
-    ck_assert(isAddressEq(&address, &poutput->Address));
-    ck_assert(genCoins == poutput->Coins);
-    ck_assert(genCoins == poutput->Hours);
+    BlockBody__Handle blockbody = 0;
+    result = SKY_coin_GetBlockBody(block, &blockbody);
+    ck_assert(result == SKY_OK);
+    Transactions__Handle trans = 0;
+    result = SKY_coin_BlockBody_Transactions(blockbody, &trans);
+    ck_assert(result == SKY_OK);
+    GoInt transLength;
+    result = SKY_coin_Transactions_Length(trans, &transLength);
+    ck_assert(result == SKY_OK);
+    ck_assert_int_eq(1, transLength);
+    Transaction__Handle tran = 0;
+    result = SKY_coin_Transactions_GetAt(trans, 0, &tran);
+    ck_assert(result == SKY_OK);
+    GoInt inLen;
+    SKY_coin_Transaction_GetInputsCount(tran, &inLen);
+    ck_assert_int_eq(0, inLen);
+    GoInt sigLen;
+    SKY_coin_Transaction_GetSignaturesCount(tran, &sigLen);
+    ck_assert_int_eq(0, sigLen);
+    GoInt outLen;
+    SKY_coin_Transaction_GetOutputsCount(tran, &outLen);
+    ck_assert_int_eq(1, outLen);
+    coin__TransactionOutput poutput;
+    SKY_coin_Transaction_GetOutputAt(tran, 0, &poutput);
+    ck_assert(isAddressEq(&address, &poutput.Address));
+    ck_assert(genCoins == poutput.Coins);
+    ck_assert(genCoins == poutput.Hours);
 }
 END_TEST
 
@@ -241,7 +290,7 @@ START_TEST(TestCreateUnspent)
     cipher__SHA256 hash;
     coin__Transaction* ptx;
     Transaction__Handle handle;
-    ptx = makeEmptyTransaction(&handle);
+    makeEmptyTransaction(&handle);
     result = SKY_coin_Transaction_PushOutput(handle, &address, 11000000, 255);
     ck_assert_msg(result == SKY_OK, "SKY_coin_Transaction_PushOutput failed");
     coin__BlockHeader bh;
@@ -266,14 +315,18 @@ START_TEST(TestCreateUnspent)
         }
         ck_assert(bh.Time == ux.Head.Time);
         ck_assert(bh.BkSeq == ux.Head.BkSeq);
+        strcpy(hash, " ");
         result = SKY_coin_Transaction_Hash(handle, &hash);
         ck_assert_msg(result == SKY_OK, "SKY_coin_Transaction_Hash failed");
         ck_assert(isU8Eq(hash, ux.Body.SrcTransaction, sizeof(cipher__SHA256)));
-        ck_assert(t[i].index < ptx->Out.len);
-        coin__TransactionOutput* poutput = (coin__TransactionOutput*)ptx->Out.data;
-        ck_assert(isAddressEq(&ux.Body.Address, &poutput->Address));
-        ck_assert(ux.Body.Coins == poutput->Coins);
-        ck_assert(ux.Body.Hours == poutput->Hours);
+        GoInt ptxLen;
+        SKY_coin_Transaction_GetOutputsCount(handle, &ptxLen);
+        ck_assert_int_lt(t[i].index, ptxLen);
+        coin__TransactionOutput poutput;
+        SKY_coin_Transaction_GetOutputAt(handle, 0, &poutput);
+        ck_assert(isAddressEq(&ux.Body.Address, &poutput.Address));
+        ck_assert(ux.Body.Coins == poutput.Coins);
+        ck_assert(ux.Body.Hours == poutput.Hours);
     }
 }
 END_TEST
@@ -289,7 +342,7 @@ START_TEST(TestCreateUnspents)
     cipher__SHA256 hash;
     coin__Transaction* ptx;
     Transaction__Handle handle;
-    ptx = makeEmptyTransaction(&handle);
+    makeEmptyTransaction(&handle);
     result = SKY_coin_Transaction_PushOutput(handle, &address, 11000000, 255);
     ck_assert_msg(result == SKY_OK, "SKY_coin_Transaction_PushOutput failed");
     coin__BlockHeader bh;
@@ -301,21 +354,22 @@ START_TEST(TestCreateUnspents)
     result = SKY_coin_CreateUnspents(&bh, handle, &uxs);
     ck_assert_msg(result == SKY_OK, "SKY_coin_CreateUnspents failed");
     registerMemCleanup(uxs.data);
-    ck_assert(uxs.len == 1);
-    ck_assert_int_eq(uxs.len, ptx->Out.len);
+    GoInt ptxLen;
+    SKY_coin_Transaction_GetOutputsCount(handle, &ptxLen);
+    ck_assert_int_eq(uxs.len, ptxLen);
     coin__UxOut* pout = (coin__UxOut*)uxs.data;
-    coin__TransactionOutput* ptxout = (coin__TransactionOutput*)ptx->Out.data;
     for (int i = 0; i < uxs.len; i++) {
+        coin__TransactionOutput ptxout;
+        SKY_coin_Transaction_GetOutputAt(handle, i, &ptxout);
         ck_assert(bh.Time == pout->Head.Time);
         ck_assert(bh.BkSeq == pout->Head.BkSeq);
         result = SKY_coin_Transaction_Hash(handle, &hash);
         ck_assert_msg(result == SKY_OK, "SKY_coin_Transaction_Hash failed");
         ck_assert(isU8Eq(hash, pout->Body.SrcTransaction, sizeof(cipher__SHA256)));
-        ck_assert(isAddressEq(&pout->Body.Address, &ptxout->Address));
-        ck_assert(pout->Body.Coins == ptxout->Coins);
-        ck_assert(pout->Body.Hours == ptxout->Hours);
+        ck_assert(isAddressEq(&pout->Body.Address, &ptxout.Address));
+        ck_assert(pout->Body.Coins == ptxout.Coins);
+        ck_assert(pout->Body.Hours == ptxout.Hours);
         pout++;
-        ptxout++;
     }
 }
 END_TEST
@@ -327,12 +381,12 @@ Suite* coin_blocks(void)
 
     tc = tcase_create("coin.block");
     tcase_add_checked_fixture(tc, setup, teardown);
-    // tcase_add_test(tc, TestNewBlock);
-    // tcase_add_test(tc, TestBlockHashHeader);
-    // tcase_add_test(tc, TestBlockHashBody);
-    // tcase_add_test(tc, TestNewGenesisBlock);
-    // tcase_add_test(tc, TestCreateUnspent);
-    tcase_add_test(tc, TestCreateUnspents);
+    tcase_add_test(tc, TestNewBlock); //ok
+    tcase_add_test(tc, TestBlockHashHeader);
+    tcase_add_test(tc, TestBlockHashBody);
+    tcase_add_test(tc, TestNewGenesisBlock); //ok
+    tcase_add_test(tc, TestCreateUnspent);   //ok
+    tcase_add_test(tc, TestCreateUnspents);  //ok
     suite_add_tcase(s, tc);
     tcase_set_timeout(tc, 150);